Use files for commit input.
I've started using multiple commit lines which wasn't possible with the
old setup. With the new method vue editor will be opened to format a
commit message. The commit method now accepts files input so it is
easily possible to only commit certain files with the default set to
commit all changes.